+++ .eslintrc.json
@@ -62,13 +62,48 @@
             "builtin",
             "external"
           ],
-          "internal",
-          "parent",
           [
+            "internal",
+            "parent",
             "sibling",
             "index"
           ],
           "object"
+        ],
+        "pathGroups": [
+          {
+            "pattern": "@app/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
+            "pattern": "@components/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
+            "pattern": "@hooks/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
+            "pattern": "@utils/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
+            "pattern": "@fonts/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
+            "pattern": "@images/**",
+            "group": "internal",
+            "position": "before"
+          }
+        ],
+        "pathGroupsExcludedImportTypes": [
+          "builtin"
         ],
         "newlines-between": "always",
         "alphabetize": {
