+++ src/app/layout.recoil.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import React from 'react';
 import { RecoilRoot } from 'recoil';
