+++ src/components/pages/TodoUserListPage.tsx
@@ -3,7 +3,7 @@
 import styled from '@emotion/styled';
 import React from 'react';
 
-import TodoListApp from '../todo-list/App';
+import TodoListApp from '@components/todo-list/App';
 
 const Container = styled.div`
   padding: 128px 32px;
