+++ src/components/todo-list/components/CheckBox.tsx
@@ -1,56 +1,18 @@
-import styled from '@emotion/styled';
 import React from 'react';
 
-import CheckIcon from '../../../../public/images/Check.svg';
+import {
+  CheckboxStyled,
+  CheckboxWrapper,
+  HiddenCheckbox,
+} from '@components/todo-list/components/CheckBox.emotion';
+
+import CheckIcon from '@images/Check.svg';
 
 interface Props {
   id: string;
   isChecked: boolean;
   handleToggle: () => void;
 }
-
-const CheckboxWrapper = styled.div`
-  position: relative;
-  width: 32px;
-  height: 32px;
-`;
-
-const HiddenCheckbox = styled.input`
-  position: absolute;
-  opacity: 0;
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  cursor: pointer;
-`;
-
-const StyledCheckbox = styled.div<{ checked: boolean }>`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  border: 1px solid
-    ${({ checked, theme: { colors } }) =>
-      checked ? colors.primary200 : colors.border};
-  border-radius: 50%;
-  background-color: ${({ checked, theme: { colors } }) =>
-    checked ? colors.primary100 : colors.white};
-  cursor: pointer;
-  transition:
-    background-color 0.2s,
-    border-color 0.2s;
-
-  svg {
-    width: 20px;
-    height: 20px;
-  }
-
-  svg path {
-    transform: scale(0.83);
-    fill: ${({ theme: { colors } }) => colors.primary200};
-  }
-`;
 
 export default function CheckBox({ id, isChecked, handleToggle }: Props) {
   return (
@@ -63,9 +25,9 @@ export default function CheckBox({ id, isChecked, handleToggle }: Props) {
         aria-checked={isChecked}
         aria-label="Toggle Todo"
       />
-      <StyledCheckbox checked={isChecked}>
+      <CheckboxStyled checked={isChecked}>
         {isChecked && <CheckIcon />}
-      </StyledCheckbox>
+      </CheckboxStyled>
     </CheckboxWrapper>
   );
 }
