+++ src/components/todo-list/components/TodoList.emotion.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const TodoListContainer = styled.ul`
+  list-style: none;
+  padding: 0;
+`;
