+++ src/components/todo-list/components/TodoListCount.emotion.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const TodoListCountContainer = styled.div`
+  margin-top: 32px;
+  padding: 16px 0 16px 16px;
+  font-size: 20px;
+  font-weight: 400;
+  line-height: 28px;
+`;
