+++ src/components/todo-list/components/TodoListCount.tsx
@@ -1,22 +1,15 @@
-import styled from "@emotion/styled";
-import React from "react";
+import React from 'react';
 
-interface FooterProps {
+import { TodoListCountContainer } from '@components/todo-list/components/TodoListCount.emotion';
+
+interface TodoListCountProps {
   totalCount: number;
 }
 
-const Container = styled.div`
-  margin-top: 32px;
-  padding: 16px 0 16px 16px;
-  font-size: 20px;
-  font-weight: 400;
-  line-height: 28px;
-`;
-
-export default function TodoListCount({totalCount}: FooterProps) {
+export default function TodoListCount({ totalCount }: TodoListCountProps) {
   return (
-    <Container>
+    <TodoListCountContainer>
       <p>총 {totalCount}개</p>
-    </Container>
+    </TodoListCountContainer>
   );
-};
+}
