+++ src/components/todo-list/components/TodoMenu.tsx
@@ -1,7 +1,10 @@
-import styled from '@emotion/styled';
 import React from 'react';
 
-import { Filter } from '../types/types';
+import {
+  Button,
+  TodoMenuContainer,
+} from '@components/todo-list/components/TodoMenu.emotion';
+import { Filter } from '@components/todo-list/types/types';
 
 interface TodoMenuProps {
   options: Filter[];
@@ -9,38 +12,13 @@ interface TodoMenuProps {
   handleChangeFilter: (filter: Filter) => void;
 }
 
-const Container = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-const Button = styled.button<{ active: boolean }>`
-  width: 108px;
-  height: 40px;
-  border-radius: 12px;
-  background-color: ${({ active, theme: { colors } }) =>
-    active ? colors.primary100 : colors.white};
-  color: ${({ active, theme: { colors } }) =>
-    active ? colors.primary200 : colors.textTertiary};
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 40px;
-  cursor: pointer;
-
-  &:hover {
-    background-color: ${({ theme: { colors } }) => colors.primary100};
-    color: ${({ theme: { colors } }) => colors.primary200};
-  }
-`;
-
 export default function TodoMenu({
   options,
   currentFilter,
   handleChangeFilter,
 }: TodoMenuProps) {
   return (
-    <Container>
+    <TodoMenuContainer>
       {options.map(option => (
         <Button
           key={option}
@@ -50,6 +28,6 @@ export default function TodoMenu({
           {option.toUpperCase()}
         </Button>
       ))}
-    </Container>
+    </TodoMenuContainer>
   );
 }
